bcmdhd: workaround for driver compiled as a module
BCMDHD driver on several Samsung and few Sony devices is compiled as a
module rather than being built into the kernel. Hence, the wlan0 interface
is removed when the WiFi is turned off and the driver is unloaded. This
causes the WiFi HAL to crash when the device is booted with WiFi turned off
since with the absence of wlan0 interface, the WiFi interfaces in the HAL
are not initialized.

As a workaround, we check if the driver is loaded and if not, we temporarily load
and unload the WiFi driver inorder to query the interfaces and initialize them.
This atleast prevents the system from crashing and rebooting when the device is
suspended.

complemets: https://gerrit.omnirom.org/11024

Change-Id: Ia7c2ce79a58e9eb6711fbf5f53f1d449cc27af87